Create casting
This is a journal entry regarding the subject of casting. This entry demonstrates how to assign values to multiple variables, display variable types, and change variable types at will. There are some circumstances in which we may want to change a variable's type.
A common usage of casting is altering the variable type of user input.